requirements.txt
# ==================== Core Framework ====================
fastapi==0.115.6
uvicorn[standard]==0.34.0
starlette==0.41.3

# ==================== Database ====================
# PRODUCTION NOTE: Use async drivers for non-blocking I/O
SQLAlchemy==2.0.36
sqlmodel==0.0.27
alembic==1.14.0
greenlet==3.2.4
# Async SQLite driver (dev/test)
aiosqlite==0.21.0
# PostgreSQL drivers
psycopg2-binary==2.9.10  # Sync operations (migrations, init_db)
psycopg==3.2.3  # Async operations (API endpoints) - Python 3.14 compatible

# ==================== Authentication & Security ====================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
cryptography==46.0.3
python-multipart==0.0.20

# ==================== Configuration ====================
pydantic==2.12.4
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.0.1

# ==================== HTTP Client ====================
httpx==0.28.1
httpcore==1.0.9
certifi==2025.10.5

# ==================== Validation ====================
email_validator==2.2.0

# ==================== Development Tools ====================
# Type checking
mypy==1.18.2
mypy_extensions==1.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.2

# ==================== Utilities ====================
# Date/time
python-dateutil==2.9.0

# PDF processing
PyPDF2==3.0.1
reportlab==4.2.5

# Image processing
pillow==12.0.0

# Rate limiting
slowapi==0.1.9
limits==5.6.0

# Monitoring (optional but recommended)
sentry-sdk==2.42.0

# CLI tools
typer==0.19.2
click==8.3.0
rich==13.9.4
shellingham==1.5.4

# Template engine (for emails)
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3

# Markdown
markdown-it-py==4.0.0
mdurl==0.1.2

# Additional dependencies
annotated-types==0.7.0
anyio==4.11.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pyasn1==0.6.1
pycparser==2.23
cffi==2.0.0
Pygments==2.19.2
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
chardet==5.2.0
charset-normalizer==3.4.4
Deprecated==1.2.18
dnspython==2.8.0
ecdsa==0.19.1
rsa==4.9.1
PyYAML==6.0.3
requests==2.32.3
rich-toolkit==0.15.1
rignore==0.7.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
